hw: vfio: Move quirks into their own directory
For configurability purposes, we're going to split pci-quirks into
smaller pieces. This is more cleanly done under its own sub-directory.
