+++ Makefile.objs
@@ -237,6 +237,7 @@ trace-events-subdirs += hw/timer
 trace-events-subdirs += hw/tpm
 trace-events-subdirs += hw/usb
 trace-events-subdirs += hw/vfio
+trace-events-subdirs += hw/vfio/quirks
 trace-events-subdirs += hw/virtio
 trace-events-subdirs += hw/xen
 trace-events-subdirs += io
