+++ hw/vfio/quirks/Makefile.objs
@@ -1,0 +1,1 @@
+obj-$(CONFIG_PCI) += pci.o
